backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
cryptography>=42.0.0
google-genai>=1.0.0
apscheduler>=3.10.0
slowapi>=0.1.9
python-dotenv>=1.0.0
email-validator>=2.1.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.115.0
slack-sdk>=3.26.0
notion-client>=2.2.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0
